+++ src/game/loot/index.ts
@@ -2,3 +2,6 @@ export * from './LootTypes';
 export * from './LootRegistry';
 export * from './LootGenerator';
 export * from './EquipmentStatAggregator';
+export * from './InventoryRuntime';
+export * from './GroundLootTypes';
+export * from './GroundLootRuntime';